requirements.txt
Flask
Werkzeug
Jinja2
flask-sqlalchemy
sqlalchemy

google-api-python-client
google-auth
google-auth-oauthlib
oauth2client
google-cloud-vision
google-generativeai
requests
gunicorn
google-generativeai
docker
requests
python-dotenv
psutil
reportlab
openpyxl
beautifulsoup4
sentence-transformers
scikit-learn
python-docx
celery
redis